requirements.txt
# Core Agent Framework
deepagents==0.4.1
langchain==1.2.10
langchain-community==0.4.1
langchain-openai==1.1.9
langgraph==1.0.8

# LLM & Embedding
llama-index==0.14.14
llama-index-llms-openai==0.6.18
llama-index-embeddings-openai==0.5.1
llama-index-vector-stores-postgres==0.7.3

# Database & Storage
sqlalchemy==2.0.46
psycopg[binary]==3.3.2
asyncpg==0.31.0
pgvector==0.4.2

# Knowledge Graph & Memory
graphiti-core[falkordb]==0.27.1

# Observability & Tracing
langfuse==3.14.1
openinference-instrumentation-llama-index==4.3.9

# Utilities
python-dotenv==1.2.1
tiktoken==0.12.0
python-pptx==1.0.2

# Testing
pytest==9.0.2
pytest-asyncio==1.3.0
